split api into redirect-server and verification-server
