+++ api/lib/OAuthService/OAuthService.js
@@ -1,4 +1,5 @@
 const moment = require('moment');
+const axios = require('axios');
 
 const HARVEST_API_OAUTH_URL = process.env.HARVEST_API_OAUTH_URL;
 const OAUTH_APP_ID = process.env.OAUTH_APP_ID;
